refactor: move register.py into src/pipeline/
Fits alongside other pipeline orchestration modules (train, evaluate,
score, finalize). Invoked via uv run -m src.pipeline.register.
