+++ .github/workflows/run-training.yml
@@ -141,7 +141,7 @@ jobs:
           --set-env-vars GCP_PROJECT_ID=${{ env.GCP_PROJECT_ID }} \
           --set-env-vars ENVIRONMENT=${{ needs.setup.outputs.env_tag }} \
           --command="bash" \
-          --args="-c,uv run python -m src.utils.sync_experiments --download && uv run register.py"
+          --args="-c,uv run python -m src.utils.sync_experiments --download && uv run -m src.pipeline.register"
         
         # Execute with streaming logs and wait for completion
         gcloud run jobs execute ${JOB_NAME} \
